Remove second motor and DIO Constants class
Turns out, thanks to a misunderstanding, we didn't need the second motor! Woo! I removed that. Also, I added a class in the Constants file for DIO ports (of which there should be 4 total, but only 1 here).

One more change, intake speeds don't need to be converted from Celsius to Fahrenheit. That's for temperature. Not sure how I made that mistake.