Add custom CTRE Utils (#76)
* Basic start

Created CTREUtils.java. Currently only has a configureDefaultCurrentLimit method, which has temporary values. I would like to know what other methods we want and what the default current limit values should be.

* Removed setting motor config in the method

* Add custom config method and added timeout value

Created a method to custom create this. Also expanded on the TODO at the top of the file.

Timeout value of 0 prevents errors from being thrown if the current limit isn't created in time. Without specifying it, this value is 50 (milliseconds).

* Added Javadocs

* Misunderstanding: fixed defaultCurrentLimit and removed createCurrentLimit

* Add default comment

Co-authored-by: Zachary Orr <516458+ZachOrr@users.noreply.github.com>