Encoder Climber
This branch has one main idea, that there is a linear relationship between encoder ticks and the string pot, and that they can be calibrated for the bottom, so that the climber "knows" what encoder tick you start from and can run from that. Also, started Clint's idea for a starting climber sequence.